fix memory
